requirements.txt
# Atlas-NYC Production Dependencies for Render
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate Limiting & Performance
slowapi==0.1.9

# Database
psycopg2-binary==2.9.9

# Google Maps
googlemaps==4.10.0

# File handling
requests==2.31.0
Pillow==10.1.0
aiofiles==23.2.1

# Validation
email-validator==2.1.0

# Monitoring
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2